refactor: categoryController 삭제 -> QuestionController 병합
