+++ src/main/java/com/pooli/question/controller/QuestionController.java
@@ -2,6 +2,8 @@ package com.pooli.question.controller;
 
 import java.util.List;
 
+import com.pooli.question.domain.dto.response.QuestionCategoryListResDto;
+import com.pooli.question.domain.dto.response.QuestionCategoryResDto;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -92,5 +94,26 @@ public class QuestionController {
 		return ResponseEntity.ok(
 			    QuestionResDto.builder().build()
 			);
+	}
+
+	@Operation(
+			summary = "문의 카테고리 목록 조회",
+			description = "관리자 및 유저가 문의 카테고리 전체 목록을 조회한다. 삭제 여부는 deletedAt 값으로 구분한다.")
+	@ApiResponses({
+			@ApiResponse(responseCode = "200", description = "조회 성공"),
+			@ApiResponse(responseCode = "404", description = "카테고리 정보가 존재하지 않음"),
+			@ApiResponse(responseCode = "500", description = "서버 오류")
+	})
+	@GetMapping("/category")
+	public ResponseEntity<QuestionCategoryListResDto> getQuestionCategories() {
+		QuestionCategoryResDto questionCategoryResDto = QuestionCategoryResDto.builder()
+				.questionCategoryId(1)
+				.questionCategoryName("요금제 문의")
+				.build();
+
+		QuestionCategoryListResDto questionCategoryListResDto = QuestionCategoryListResDto.builder()
+				.questionCategories(List.of(questionCategoryResDto))
+				.build();
+		return ResponseEntity.ok(questionCategoryListResDto);
 	}
 }
